refactor(ui): consolidate and reorganise machine/pod utils
