+++ ui/src/app/kvm/views/KVMList/KVMListTable/TypeColumn/TypeColumn.tsx
@@ -1,8 +1,8 @@
 import { useSelector } from "react-redux";
 
 import DoubleRow from "app/base/components/DoubleRow";
-import { formatHostType } from "app/kvm/utils";
 import podSelectors from "app/store/pod/selectors";
+import { formatHostType } from "app/store/pod/utils";
 import type { RootState } from "app/store/root/types";
 
 type Props = { id: number };
