--- ui/src/app/machines/views/MachineDetails/MachineStorage/types.ts
@@ -1,38 +1,0 @@
-import type { Filesystem } from "app/store/machine/types";
-
-export type NormalisedFilesystem = {
-  actions: string[];
-  fstype: Filesystem["fstype"];
-  id: Filesystem["id"];
-  mountOptions: Filesystem["mount_options"];
-  mountPoint: Filesystem["mount_point"];
-  name: string | null;
-  parentId: number | null;
-  parentType: string | null;
-  size: number | null;
-};
-
-export type NormalisedStorageDevice = {
-  actions: string[];
-  boot: boolean | null;
-  firmware: string | null;
-  id: number;
-  model: string | null;
-  name: string;
-  numaNodes: number[];
-  parentType: string | null;
-  serial: string | null;
-  size: number;
-  tags: string[];
-  testStatus: number | null;
-  type: string;
-  usedFor: string;
-};
-
-export type SeparatedDiskData = {
-  available: NormalisedStorageDevice[];
-  cacheSets: NormalisedStorageDevice[];
-  datastores: NormalisedFilesystem[];
-  filesystems: NormalisedFilesystem[];
-  used: NormalisedStorageDevice[];
-};
