+++ ui/src/app/rsd/views/RSDList/RSDListHeader/RSDListHeader.tsx
@@ -7,10 +7,10 @@ import { Link, useLocation } from "react-router-dom";
 import RSDListActionMenu from "./RSDListActionMenu";
 
 import SectionHeader from "app/base/components/SectionHeader";
-import { getVMHostCount } from "app/kvm/utils";
 import RSDActionFormWrapper from "app/rsd/components/RSDActionFormWrapper";
 import { actions as podActions } from "app/store/pod";
 import podSelectors from "app/store/pod/selectors";
+import { getVMHostCount } from "app/store/pod/utils";
 
 const RSDListHeader = (): JSX.Element => {
   const dispatch = useDispatch();
