+++ ui/src/app/store/machine/utils/hooks.ts
@@ -1,0 +1,138 @@
+import { useEffect } from "react";
+
+import { useDispatch, useSelector } from "react-redux";
+
+import { isMachineStorageConfigurable } from "./storage";
+
+import { general as generalActions } from "app/base/actions";
+import generalSelectors from "app/store/general/selectors";
+import type { Machine } from "app/store/machine/types";
+import type { RootState } from "app/store/root/types";
+import type { Host } from "app/store/types/host";
+import { NodeStatus } from "app/store/types/node";
+
+/**
+ * Check if a machine can be edited.
+ * @param machine - A machine object.
+ * @param ignoreRackControllerConnection - Whether the editable check should
+ *                                         include whether the rack controller
+ *                                          is connected.
+ * @returns Whether the machine can be edited.
+ */
+export const useCanEdit = (
+  machine?: Machine | null,
+  ignoreRackControllerConnection = false
+): boolean => {
+  const isRackControllerConnected = useIsRackControllerConnected();
+  if (!machine) {
+    return false;
+  }
+  return (
+    machine.permissions.includes("edit") &&
+    !machine.locked &&
+    (ignoreRackControllerConnection || isRackControllerConnected)
+  );
+};
+
+/**
+ * Check whether a machine's storage can be edited based on permissions and the
+ * machine's status.
+ * @param machine - A machine object.
+ * @returns Whether the machine's storage can be edited.
+ */
+export const useCanEditStorage = (machine: Machine | null): boolean => {
+  const canEdit = useCanEdit(machine);
+  const machineStorageConfigurable = isMachineStorageConfigurable(machine);
+  return canEdit && machineStorageConfigurable;
+};
+
+/**
+ * Format a node's OS and release into human-readable text.
+ * @param node - A node object.
+ * @returns Formatted OS string.
+ */
+export const useFormattedOS = (node?: Host | null): string => {
+  const dispatch = useDispatch();
+  const osReleases = useSelector((state: RootState) =>
+    generalSelectors.osInfo.getOsReleases(state, node?.osystem)
+  );
+
+  useEffect(() => {
+    dispatch(generalActions.fetchOsInfo());
+  }, [dispatch]);
+
+  if (!node || !node.osystem || !node.distro_series) {
+    return "";
+  }
+
+  const machineRelease = osReleases.find(
+    (release) => release.value === node.distro_series
+  );
+  if (machineRelease) {
+    return node.osystem === "ubuntu"
+      ? machineRelease.label.split('"')[0].trim()
+      : machineRelease.label;
+  }
+  return `${node.osystem}/${node.distro_series}`;
+};
+
+/**
+ * Check if a machine has an invalid architecture.
+ * @param machine - A machine object.
+ * @returns Whether the machine has an invalid architecture.
+ */
+export const useHasInvalidArchitecture = (
+  machine?: Machine | null
+): boolean => {
+  const dispatch = useDispatch();
+  const architectures = useSelector(generalSelectors.architectures.get);
+
+  useEffect(() => {
+    dispatch(generalActions.fetchArchitectures());
+  }, [dispatch]);
+
+  if (!machine) {
+    return false;
+  }
+
+  return (
+    machine.architecture === "" || !architectures.includes(machine.architecture)
+  );
+};
+
+/**
+ * Check if the networking information can be edited.
+ * @return Wether networking is disabled.
+ */
+export const useIsAllNetworkingDisabled = (
+  machine?: Machine | null
+): boolean => {
+  const canEdit = useCanEdit(machine, true);
+  return (
+    !canEdit ||
+    !machine ||
+    ![
+      NodeStatus.NEW,
+      NodeStatus.READY,
+      NodeStatus.FAILED_TESTING,
+      NodeStatus.ALLOCATED,
+      NodeStatus.BROKEN,
+    ].includes(machine.status)
+  );
+};
+
+/**
+ * Check if the rack controller is connected.
+ * @returns Whether the rack controller is connected.
+ */
+export const useIsRackControllerConnected = (): boolean => {
+  const dispatch = useDispatch();
+  const powerTypes = useSelector(generalSelectors.powerTypes.get);
+
+  useEffect(() => {
+    dispatch(generalActions.fetchPowerTypes());
+  }, [dispatch]);
+
+  // If power types exist then a rack controller is connected.
+  return powerTypes.length > 0;
+};
